Avoid a warning when compiling by casting off_t to size_t.

+++ compat_sha2_h.c
@@ -992,7 +992,7 @@ SHA256FileChunk(const char *filename, char *buf, off_t off, off_t len)
 		return (NULL);
 	}
 
-	while ((nr = read(fd, buffer, MINIMUM(sizeof(buffer), len))) > 0) {
+	while ((nr = read(fd, buffer, MINIMUM(sizeof(buffer), (size_t)len))) > 0) {
 		SHA256Update(&ctx, buffer, nr);
 		if (len > 0 && (len -= nr) == 0)
 			break;
@@ -1039,7 +1039,7 @@ SHA384FileChunk(const char *filename, char *buf, off_t off, off_t len)
 		return (NULL);
 	}
 
-	while ((nr = read(fd, buffer, MINIMUM(sizeof(buffer), len))) > 0) {
+	while ((nr = read(fd, buffer, MINIMUM(sizeof(buffer), (size_t)len))) > 0) {
 		SHA384Update(&ctx, buffer, nr);
 		if (len > 0 && (len -= nr) == 0)
 			break;
@@ -1086,7 +1086,7 @@ SHA512FileChunk(const char *filename, char *buf, off_t off, off_t len)
 		return (NULL);
 	}
 
-	while ((nr = read(fd, buffer, MINIMUM(sizeof(buffer), len))) > 0) {
+	while ((nr = read(fd, buffer, MINIMUM(sizeof(buffer), (size_t)len))) > 0) {
 		SHA512Update(&ctx, buffer, nr);
 		if (len > 0 && (len -= nr) == 0)
 			break;
